Refactor and clean components
